+++ examples/rendering_plane.py
@@ -1,7 +1,7 @@
 import napari
 from skimage import data
 
-from napari_threedee.manipulator.manipulator import RenderPlaneTranslator
+from napari_threedee.manipulators import RenderPlaneManipulator
 
 viewer = napari.Viewer(ndisplay=3)
 
@@ -17,7 +17,7 @@ volume_layer = viewer.add_image(
 # plane should be in 'additive' blending mode or depth looks all wrong
 plane_parameters = {
     'position': (32, 32, 32),
-    'normal': (0, 1, 0),
+    'normal': (1, 0, 0),
     'thickness': 10,
     'enabled': True
 }
@@ -31,6 +31,6 @@ plane_layer = viewer.add_image(
     experimental_slicing_plane=plane_parameters
 )
 
-axis = RenderPlaneTranslator(viewer=viewer, layer=plane_layer, line_length=10)
+axis = RenderPlaneManipulator(viewer=viewer, layer=plane_layer, line_length=10)
 
 napari.run()
