+++ src/napari_threedee/manipulators/__init__.py
@@ -1,0 +1,2 @@
+from .layer_manipulator import LayerManipulator
+from .render_plane_manipulator import RenderPlaneManipulator
